+++ src/api/entities-to-create/conversation.entity.ts
@@ -12,6 +12,7 @@ import { ConnectedCardEntity } from './connected-card.entity';
 import { GroupEntity } from './group.entity';
 import { JoinedConversation } from './joined-conversation.entity';
 import { MessageEntity } from './message.entity';
+import { NotificationEntity } from "./notification.entity";
 
 @Entity({ name: 'conversation' })
 export class ConversationEntity {
@@ -39,6 +40,9 @@ export class ConversationEntity {
 
   @OneToMany(() => JoinedConversation, joinedConversation => joinedConversation.conversation)
   joinedProfiles: JoinedConversation[];
+
+  @OneToMany(() => NotificationEntity, notification => notification.conversation)
+  notifications: NotificationEntity[];
 
   // ______________________________________________________
   // Timestamps