+++ src/api/user/domain/dto/user-login.dto.ts
@@ -1,0 +1,3 @@
+export class UserLoginDto {
+  constructor(public readonly login: string, public readonly password: string) {}
+}