+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [0.0.4](https://github.com/WalletShareOrg/api-wallet-share/tree/V_0.0.4) - 2023-03-10
+
+### Features
+- add referral code entity
+- add notification entity
+- update README.md
+
 # [0.0.3](https://github.com/WalletShareOrg/api-wallet-share/tree/V_0.0.3) - 2023-03-10
 
 ### Features